refactor: delete card shoe serv and main page
