+++ YouTube-client-app/src/app/youtube/pages/cards-list-page/card-item/card-item.component.scss
