+++ YouTube-client-app/src/app/youtube/pages/cards-list-page/card-item/card-item.component.ts
@@ -2,10 +2,10 @@ import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { ButtonComponent } from '../../../../../shared/components/button/button.component';
-import { CardStatsComponent } from '../../../../components/card-statistic/card-stats.component';
-import { CardColorDirective } from '../../../../directives/card-color.directive';
-import { CardItemModel } from '../../../../models/card-item.model';
+import { ButtonComponent } from '../../../../shared/components/button/button.component';
+import { CardStatsComponent } from '../../../components/card-statistic/card-stats.component';
+import { CardColorDirective } from '../../../directives/card-color.directive';
+import { CardItemModel } from '../../../models/card-item.model';
 
 @Component({
   selector: 'app-card-item',
