+++ YouTube-client-app/src/app/youtube/pages/cards-list-page/cards-list.component.scss
