--- YouTube-client-app/src/app/youtube/pages/main-page/main-page.component.html
@@ -1,4 +1,0 @@
-<main class="main-page">
-  <app-cards-list
-  ></app-cards-list>
-</main>
