+++ YouTube-client-app/src/app/youtube/youtube-routing.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { CardsListComponent } from './pages/cards-list-page/cards-list.component';
 // import { DetailedPageComponent } from './pages/detailed-page/detailed-page.component';
-import { MainPageComponent } from './pages/main-page/main-page.component';
+// import { MainPageComponent } from './pages/main-page/main-page.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: MainPageComponent,
+    component: CardsListComponent,
   },
   // {
   //   path: 'details/:id',
